Move chat notification files into subdirectories
