+++ src/components/Chat/NotificationMessages/NameChangedMessage.js
