+++ src/components/Chat/NotificationMessages/index.css
@@ -1,0 +1,1 @@
+@import "./UserNotificationMessage.css";
