+++ src/components/Chat/index.css
@@ -1,8 +1,8 @@
 @import "./ChatMessages.css";
 @import "./Input/index.css";
 @import "./Markup/index.css";
+@import "./NotificationMessages/index.css";
 @import "./LogMessage.css";
-@import "./UserNotificationMessage.css";
 @import "./Message.css";
 @import "./Motd.css";
